// packages/extension-ui/src/util/types.ts
export interface Auction {
  auctionCounter: number;
  auctionInfo: [string, string];
  blockchain: string;
  crowdloans: Crowdloan[];
  currentBlockNumber: number;
  minContribution: string;
  token: string;
  winning: string[];
}

export interface Crowdloan {
  fund: Fund;
  identity: DeriveAccountRegistration;
}

interface Fund {
  depositor: string;
  verifier: string | null;
  deposit: string;
  raised: string;
  end: number;
  cap: string;
  lastContribution: { ending: bigint };
  firstPeriod: number;
  lastPeriod: number;
  trieIndex: number;
  paraId: string;
  hasLeased: boolean;
  contributionBlock?: number;
  contributionTimestamp?: number;
  unlockingBlock?: number;
}